# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger new-api",
        "contact": {}
    },
    "paths": {
...
        "/api/custom-oauth-provider/": {
            "get": {
                "tags": [
                    "OAuth 提供商"
                ],
                "summary": "返回所有的自定义 OAuth 提供商",
                "responses": {}
            },
            "put": {
                "tags": [
                    "OAuth 提供商"
                ],
                "summary": "更新一个 OAuth 提供商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OAuth提供商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "post": {
                "tags": [
                    "OAuth 提供商"
                ],
                "summary": "创建一个 OAuth 提供商",
                "responses": {}
            }
        },
        "/api/custom-oauth-provider/:id": {
            "get": {
                "tags": [
                    "OAuth 提供商"
                ],
                "summary": "返回给定的 OAuth 提供商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OAuth提供商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "OAuth 提供商"
                ],
                "summary": "删除一个 OAuth 提供商",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "OAuth提供商ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/oauth/:provider": {
            "get": {
                "tags": [
                    "通用"
                ],
                "summary": "处理所有标准 OAuth 提供程序的 OAuth 回调",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth 提供商",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/user/:id/oauth/bindings": {
            "get": {
                "tags": [
                    "管理员"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/:id/oauth/bindings/:provider_id": {
            "delete": {
                "tags": [
                    "管理员"
                ],
                "summary": "解绑给定的 OAuth 供应商",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OAuth供应商ID",
                        "name": "provider_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/user/oauth/bindings/:provider_id": {
            "post": {
                "tags": [
                    "已登录用户"
                ],
                "summary": "解绑指定的 OAuth",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth供应商",
                        "name": "provider_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    }
}